添加相机控制等功能。形成一个比较完备的软光栅渲染器

+++ src/Vec.h
@@ -1,10 +1,7 @@
 ﻿#pragma once
-#include <cfloat>
-#include <cmath>
-#include <vector>
 #include <iostream>
-#include <array>
-constexpr double MY_PI = 3.14159;
+
+constexpr float MY_PI = 3.14159f;
 
 template <class T, size_t N>
 class Vec;
